Some third party app stores have started publishing my builds of the app, this is not good since it will lead to users using old builds that might not work properly anymore.
When people build the app themselves they are aware that they will need to rebuild it later, people downloading from third party stores will no be so they will end up using older versions that will break and lead to frustration with Booksonic.

There is also the issue that some parts of Booksonic is relying on third party APIs, when people build the app themselves they are providing their own keys, when they are using my build they are using my keys meaning I will have to cover the cost.

For now we encourage users to get it from Play or build it themselves by only giving access to the demo server if it was signed with popeens key and then sideloaded. Anyone else building it will not be locked down to demo.